Initial work to support s3ql 3.x

+++ src/fsck.py
@@ -11,9 +11,8 @@ from . import CURRENT_FS_REV, BUFSIZE, CTRL_INODE, ROOT_INODE
 from .backends.common import NoSuchObject
 from .backends.comprenc import ComprencBackend
 from .backends.local import Backend as LocalBackend
-from .common import (inode_for_path, sha256_fh, get_path, get_backend_cachedir,
-                     get_seq_no, is_mounted, get_backend, load_params,
-                     save_params, time_ns)
+from .common import (inode_for_path, sha256_fh, get_path, get_seq_no, is_mounted,
+                     get_backend, load_params, save_params, time_ns)
 from .database import NoSuchRowError, Connection
 from .metadata import create_tables, dump_and_upload_metadata, download_metadata
 from .parse_args import ArgumentParser
@@ -58,8 +57,12 @@ class Fsck(object):
         # don't move them there repeatedly)
         self.moved_inodes = set()
 
-    def check(self):
+    def check(self, check_cache=True):
         """Check file system
+
+        If *check_cache* is False, assume that all cache files are clean (aka
+        have been uploaded to the backend). If *check_cache* is ``keep``, do not
+        remove cache files on exit.
 
         Sets instance variable `found_errors`.
         """
@@ -75,7 +78,9 @@ class Fsck(object):
         self.conn.execute('CREATE INDEX tmp5 ON ext_attributes(name_id)')
         try:
             self.check_lof()
-            self.check_cache()
+            self.check_uploads()
+            if check_cache:
+                self.check_cache(check_cache == 'keep')
             self.check_names_refcount()
 
             self.check_contents_name()
@@ -152,15 +157,42 @@ class Fsck(object):
                     self.uncorrectable_errors = True
 
 
-    def check_cache(self):
+    def check_uploads(self):
+        '''Drop rows for objects that were never successfully uploaded'''
+
+        for (obj_id,) in self.conn.query('SELECT id FROM objects WHERE size == -1'):
+            self.log_error('Cleaning up interrupted upload of object %s', obj_id)
+            self.found_errors = True
+
+            # If there are affected files, they'll be picked up by
+            # check_inode_blocks_block_id().
+            self.conn.execute('DELETE FROM blocks WHERE obj_id = ?', (obj_id,))
+
+        self.conn.execute('DELETE FROM objects WHERE size = -1')
+
+
+    def check_cache(self, keep_cache=False):
         """Commit uncommitted cache files"""
 
-        log.info("Checking cached objects...")
+        log.info("Checking for dirty cache objects...")
         if not os.path.exists(self.cachedir):
             return
+        candidates = os.listdir(self.cachedir)
 
-        for filename in os.listdir(self.cachedir):
-            self.found_errors = True
+        if sys.stdout.isatty():
+            stamp1 = 0
+        else:
+            stamp1 = float('inf')
+
+        total = len(candidates)
+        for (i, filename) in enumerate(candidates):
+            i += 1 # start at 1
+            stamp2 = time.time()
+            if stamp2 - stamp1 > 1 or i == total:
+                sys.stdout.write('\r..processed %d/%d files (%d%%)..'
+                                 % (i, total, i/total*100))
+                sys.stdout.flush()
+                stamp1 = stamp2
 
             match = re.match('^(\\d+)-(\\d+)$', filename)
             if match:
@@ -169,11 +201,31 @@ class Fsck(object):
             else:
                 raise RuntimeError('Strange file in cache directory: %s' % filename)
 
-            self.log_error("Committing block %d of inode %d to backend", blockno, inode)
-
+            # Calculate block checksum
             with open(os.path.join(self.cachedir, filename), "rb") as fh:
                 size = os.fstat(fh.fileno()).st_size
-                hash_ = sha256_fh(fh)
+                hash_should = sha256_fh(fh)
+            log.debug('%s has checksum %s', filename, hash_should)
+
+            # Check if stored block has same checksum
+            try:
+                block_id = self.conn.get_val('SELECT block_id FROM inode_blocks '
+                                             'WHERE inode=? AND blockno=?',
+                                             (inode, blockno,))
+                hash_is = self.conn.get_val('SELECT hash FROM blocks WHERE id=?',
+                                            (block_id,))
+            except NoSuchRowError:
+                hash_is = None
+            log.debug('Inode %d, block %d has checksum %s', inode, blockno,
+                      hash_is)
+            if hash_should == hash_is:
+                if not keep_cache:
+                    os.unlink(os.path.join(self.cachedir, filename))
+                continue
+
+            self.found_errors = True
+            self.log_error("Writing dirty block %d of inode %d to backend", blockno, inode)
+            hash_ = hash_should
 
             try:
                 (block_id, obj_id) = self.conn.get_row('SELECT id, obj_id FROM blocks WHERE hash=?', (hash_,))
@@ -209,7 +261,8 @@ class Fsck(object):
                 self.conn.execute('UPDATE blocks SET refcount=refcount-1 WHERE id=?', (old_block_id,))
                 self.unlinked_blocks.add(old_block_id)
 
-            os.unlink(os.path.join(self.cachedir, filename))
+            if not keep_cache:
+                os.unlink(os.path.join(self.cachedir, filename))
 
 
     def check_lof(self):
@@ -500,8 +553,8 @@ class Fsck(object):
                     continue
                 self.moved_inodes.add(inode)
 
-                affected_entries = list(self.conn.query('SELECT name, name_id, parent_inode '
-                                                        'FROM contents_v WHERE inode=?', (inode,)))
+                affected_entries = self.conn.get_list('SELECT name, name_id, parent_inode '
+                                                      'FROM contents_v WHERE inode=?', (inode,))
                 for (name, name_id, id_p) in affected_entries:
                     path = get_path(id_p, self.conn, name)
                     self.log_error("File may lack data, moved to /lost+found: %s", to_str(path))
@@ -647,8 +700,8 @@ class Fsck(object):
 
         log.info('Checking blocks (checksums)...')
 
-        for (block_id, obj_id) in list(self.conn.query('SELECT id, obj_id FROM blocks '
-                                                       'WHERE hash IS NULL')):
+        for (block_id, obj_id) in self.conn.get_list('SELECT id, obj_id FROM blocks '
+                                                     'WHERE hash IS NULL'):
             self.found_errors = True
 
             # This should only happen when there was an error during upload,
@@ -945,8 +998,8 @@ class Fsck(object):
 
                     # Copy the list, or we may pick up the same entry again and again
                     # (first from the original location, then from lost+found)
-                    affected_entries = list(self.conn.query('SELECT name, name_id, parent_inode '
-                                                            'FROM contents_v WHERE inode=?', (id_,)))
+                    affected_entries = self.conn.get_list('SELECT name, name_id, parent_inode '
+                                                            'FROM contents_v WHERE inode=?', (id_,))
                     for (name, name_id, id_p) in affected_entries:
                         path = get_path(id_p, self.conn, name)
                         self.log_error("File may lack data, moved to /lost+found: %s", to_str(path))
@@ -1040,83 +1093,6 @@ class Fsck(object):
         self.conn.execute('DELETE FROM names WHERE refcount=0 AND id=?', (name_id,))
 
 
-class ROFsck(Fsck):
-    '''
-    Check file system database only, and don't correct any errors.
-    '''
-
-    def __init__(self, path):
-
-        db = Connection(path + '.db')
-        db.execute('PRAGMA journal_mode = WAL')
-
-        param = load_params(path)
-        super().__init__(None, None, param, db)
-
-    def check(self):
-
-        self.conn.execute('BEGIN TRANSACTION')
-        try:
-            log.info('Creating temporary indices...')
-            for idx in ('tmp1', 'tmp2', 'tmp3', 'tmp4', 'tmp5'):
-                self.conn.execute('DROP INDEX IF EXISTS %s' % idx)
-            self.conn.execute('CREATE INDEX tmp1 ON blocks(obj_id)')
-            self.conn.execute('CREATE INDEX tmp2 ON inode_blocks(block_id)')
-            self.conn.execute('CREATE INDEX tmp3 ON contents(inode)')
-            self.conn.execute('CREATE INDEX tmp4 ON contents(name_id)')
-            self.conn.execute('CREATE INDEX tmp5 ON ext_attributes(name_id)')
-
-            self.check_lof()
-            self.check_names_refcount()
-
-            self.check_contents_name()
-            self.check_contents_inode()
-            self.check_contents_parent_inode()
-
-            self.check_objects_refcount()
-            self.check_objects_size()
-
-            self.check_blocks_obj_id()
-            self.check_blocks_refcount()
-            self.check_blocks_checksum()
-
-            self.check_inode_blocks_block_id()
-            self.check_inode_blocks_inode()
-
-            self.check_inodes_refcount()
-            self.check_inodes_size()
-
-            self.check_ext_attributes_name()
-            self.check_ext_attributes_inode()
-
-            self.check_symlinks_inode()
-
-            self.check_loops()
-            self.check_unix()
-            self.check_foreign_keys()
-
-        finally:
-            log.info('Dropping temporary indices...')
-            self.conn.execute('ROLLBACK')
-
-    def check_blocks_checksum(self):
-        """Check blocks.hash"""
-
-        log.info('Checking blocks (checksums)...')
-
-        for (block_id,) in self.conn.query('SELECT id FROM blocks WHERE hash IS NULL'):
-            self.found_errors = True
-            self.log_error("No cached checksum for block %d!", block_id)
-
-    def check_objects_size(self):
-        """Check objects.size"""
-
-        log.info('Checking objects (sizes)...')
-
-        for (obj_id,) in self.conn.query('SELECT id FROM objects WHERE size IS NULL'):
-            self.found_errors = True
-            self.log_error("Object %d has no size information!", obj_id)
-
 def parse_args(args):
 
     parser = ArgumentParser(
@@ -1124,17 +1100,21 @@ def parse_args(args):
 
     parser.add_log('~/.s3ql/fsck.log')
     parser.add_cachedir()
-    parser.add_authfile()
     parser.add_debug()
     parser.add_quiet()
     parser.add_backend_options()
     parser.add_version()
     parser.add_storage_url()
 
+    parser.add_argument("--keep-cache", action="store_true", default=False,
+                      help="Do not purge locally cached files on exit.")
     parser.add_argument("--batch", action="store_true", default=False,
                       help="If user input is required, exit without prompting.")
     parser.add_argument("--force", action="store_true", default=False,
                       help="Force checking even if file system is marked clean.")
+    parser.add_argument("--force-remote", action="store_true", default=False,
+                      help="Force use of remote metadata even when this would "
+                        "likely result in data loss.")
     options = parser.parse_args(args)
 
     return options
@@ -1157,32 +1137,36 @@ def main(args=None):
 
     log.info('Starting fsck of %s', options.storage_url)
 
-    cachepath = get_backend_cachedir(options.storage_url, options.cachedir)
+    cachepath = options.cachepath
     seq_no = get_seq_no(backend)
     db = None
+
+    # When there was a crash during metadata rotation, we may end up
+    # without an s3ql_metadata object.
+    meta_obj_name = 's3ql_metadata'
+    if meta_obj_name not in backend:
+        meta_obj_name += '_new'
 
     if os.path.exists(cachepath + '.params'):
         assert os.path.exists(cachepath + '.db')
         param = load_params(cachepath)
         if param['seq_no'] < seq_no:
             log.info('Ignoring locally cached metadata (outdated).')
-            param = backend.lookup('s3ql_metadata')
+            param = backend.lookup(meta_obj_name)
         else:
             log.info('Using cached metadata.')
             db = Connection(cachepath + '.db')
-            assert not os.path.exists(cachepath + '-cache') or param['needs_fsck']
 
         if param['seq_no'] > seq_no:
             log.warning('File system has not been unmounted cleanly.')
             param['needs_fsck'] = True
 
-        elif backend.lookup('s3ql_metadata')['seq_no'] != param['seq_no']:
+        elif backend.lookup(meta_obj_name)['seq_no'] != param['seq_no']:
             log.warning('Remote metadata is outdated.')
             param['needs_fsck'] = True
 
     else:
-        param = backend.lookup('s3ql_metadata')
-        assert not os.path.exists(cachepath + '-cache')
+        param = backend.lookup(meta_obj_name)
         # .db might exist if mount.s3ql is killed at exactly the right instant
         # and should just be ignored.
 
@@ -1212,17 +1196,6 @@ def main(args=None):
         param['seq_no'] = seq_no
         param['needs_fsck'] = True
 
-    if not db and os.path.exists(cachepath + '-cache'):
-        for i in itertools.count():
-            bak_name = '%s-cache.bak%d' % (cachepath, i)
-            if not os.path.exists(bak_name):
-                break
-        log.warning('Found outdated cache directory (%s), renaming to .bak%d',
-                    cachepath + '-cache', i)
-        log.warning('You should delete this directory once you are sure that '
-                    'everything is in order.')
-        os.rename(cachepath + '-cache', bak_name)
-
     if (not param['needs_fsck']
         and param['max_inode'] < 2 ** 31
         and (time.time() - param['last_fsck'])
@@ -1232,6 +1205,31 @@ def main(args=None):
         else:
             log.info('File system is marked as clean. Use --force to force checking.')
             return
+
+    # When using remote metadata, get rid of outdated local cache (so that we
+    # don't accidentally upload it)
+    outdated_cachedir = False
+    if not db:
+        try:
+            for name_ in os.listdir(cachepath + '-cache'):
+                if name_ not in ('.', '..'):
+                    outdated_cachedir = True
+                    break
+        except FileNotFoundError:
+            pass
+    if outdated_cachedir and param['needs_fsck']:
+        for i in itertools.count():
+            bak_name = '%s-cache.bak%d' % (cachepath, i)
+            if not os.path.exists(bak_name):
+                break
+        log.warning('Renaming outdated cache directory %s to .bak%d',
+                    cachepath + '-cache', i)
+        log.warning('You should delete this directory once you are sure that '
+                    'everything is in order.')
+        os.rename(cachepath + '-cache', bak_name)
+    elif outdated_cachedir:
+        log.info("Flushing outdated local cache...")
+        shutil.rmtree(cachepath + '-cache')
 
     # If using local metadata, check consistency
     if db:
@@ -1250,6 +1248,13 @@ def main(args=None):
     else:
         db = download_metadata(backend, cachepath + '.db')
 
+    # We only read cache files if the filesystem was not
+    # unmounted cleanly. On a clean unmount, the cache files can
+    # not be dirty.
+    check_cache = param['needs_fsck']
+    if check_cache and options.keep_cache:
+        check_cache = 'keep'
+
     # Increase metadata sequence no
     param['seq_no'] += 1
     param['needs_fsck'] = True
@@ -1257,14 +1262,11 @@ def main(args=None):
     save_params(cachepath, param)
 
     fsck = Fsck(cachepath + '-cache', backend, param, db)
-    fsck.check()
+    fsck.check(check_cache)
     param['max_inode'] = db.get_val('SELECT MAX(id) FROM inodes')
 
     if fsck.uncorrectable_errors:
         raise QuietError("Uncorrectable errors found, aborting.", exitcode=44+128)
-
-    if os.path.exists(cachepath + '-cache'):
-        os.rmdir(cachepath + '-cache')
 
     if param['max_inode'] >= 2 ** 31:
         renumber_inodes(db)
@@ -1307,8 +1309,8 @@ def renumber_inodes(db):
 
     create_tables(db)
     for table in ('names', 'blocks', 'objects'):
-        db.execute('DROP TABLE %s' % table)
-        db.execute('ALTER TABLE %s_old RENAME TO %s' % (table, table))
+        db.execute('INSERT INTO %s SELECT * FROM %s_old' % (table, table))
+        db.execute('DROP TABLE %s_old' % table)
 
     log.info('..mapping..')
     db.execute('CREATE TEMPORARY TABLE inode_map (rowid INTEGER PRIMARY KEY AUTOINCREMENT, id INTEGER UNIQUE)')
